Split out dovado to a component and sensor platform
